Revert "more oop"
This reverts commit eef2648893dcf5a2e70df256a5ee40ff71546cf6.
