--- src/main/java/UiCommand.java
@@ -1,7 +1,0 @@
-public class UiCommand extends Command { 
-    public UiCommand() {
-    }
-    public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
-        ui.showList(taskList);
-    }
-}
